pyproject.toml
[project]
name = "cc-fraud-pipeline"
version = "0.1.0"
description = "Real-time Credit Card Fraud Detection Pipeline (Kafka → Flink → PostgreSQL → Grafana)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cramjam>=2.11.0",
    "kafka-python-ng>=2.2.3",
    "kaggle>=2.0.1",
    "python-snappy>=0.7.3",
]
